+++ CHANGELOG.md
@@ -1,10 +1,6 @@
 # Change Log
 All notable changes to this project will be documented in this file.
- 
-## 2.0.8 2023-05-23
-### Added
-- Added Timeout functionality to PayApp TCP connection. (US344949)
-- Added token and card brand transaction id to the response for Pax. (US345636)
+
 
 ## 2.0.9 2023-06-13
 ### Added
@@ -12,7 +8,16 @@ All notable changes to this project will be documented in this file.
 - Added UPA Pay App Canada (US346223)
 - iOS SDK: C2X Refund Reversal (US359118)
 - Bug fixes and improvements.
+
+## 2.0.8 2023-05-23
+### Added
+- Added Timeout functionality to PayApp TCP connection. (US344949)
+- Added token and card brand transaction id to the response for Pax. (US345636)
  
 ### Changed
  
 ### Fixed
+
+## 2.0.10 2023-06-27
+### Added
+- Bug fixes and improvements.
