+++ Heartland-iOS-SDK-SampleApp/Heartland-iOS-SDK-SampleApp.xcodeproj/xcshareddata/xcschemes/Heartland-iOS-SDK-SampleAppTests.xcscheme
@@ -10,20 +10,8 @@
       buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
-      shouldUseLaunchSchemeArgsEnv = "YES"
-      shouldAutocreateTestPlan = "YES">
+      shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
-       <TestableReference
-            skipped = "NO"
-            parallelizable = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "DD713F852948DC3400514143"
-               BuildableName = "Heartland-iOS-SDK-SampleAppTests.xctest"
-               BlueprintName = "Heartland-iOS-SDK-SampleAppTests"
-               ReferencedContainer = "container:Heartland-iOS-SDK-SampleApp.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
