+++ Heartland-iOS-SDK.podspec
@@ -8,7 +8,7 @@
 
 Pod::Spec.new do |s|
     s.name             = "Heartland-iOS-SDK"
-    s.version          = "2.0.9"
+    s.version          = "2.0.10"
     s.summary          = "Secure Tokenized Payments by Heartland Payment Systems."
   
   # This description is used to generate tags and improve search results.
