+++ Heartland-iOS-SDK/Frameworks/GlobalPaymentsApi.xcframework/ios-arm64_x86_64-simulator/GlobalPaymentsApi.framework/Headers/GPTransaction.h
@@ -33,6 +33,7 @@
 @property (nonatomic, strong) NSString* token;
 @property (nonatomic, strong) NSString* transactionDescriptor;
 @property (nonatomic, strong) GPTransactionReference* transactionReference;
+@property (nonatomic, strong) NSString* gatewayTransactionId;
 
 + (instancetype) transactionFromId:(NSString*) transactionId;
 + (instancetype) transactionFromId:(NSString*) transactionId withPaymentMethodType:(GPPaymentMethodType) paymentMethodType;
